+++ docs/src/examples/components/Icon/Variations/IconExampleColor.shorthand.tsx
@@ -15,17 +15,29 @@ const IconExampleColor = () => (
       <Icon name="call" bordered variables={{ outline: true }} />
       <Icon name="call-video" bordered variables={{ outline: true }} />
     </div>
-    {/* The borderColor variable can change the color used for the border. */}
-    <div style={{ color: 'cornflowerblue', padding: '1rem', display: 'inline-block' }}>
-      <Icon name="calendar" bordered variables={{ borderColor: 'violet' }} />
-      <Icon name="call" bordered variables={{ borderColor: 'yellowgreen' }} />
-      <Icon name="call-video" bordered variables={{ borderColor: 'orangered' }} />
-    </div>
     {/* The color variable can change the color used for the icon, as well as the border. */}
     <div style={{ padding: '1rem', display: 'inline-block' }}>
       <Icon name="calendar" bordered variables={{ color: 'violet' }} />
       <Icon name="call" bordered variables={{ color: 'yellowgreen' }} />
       <Icon name="call-video" bordered variables={{ color: 'cornflowerblue' }} />
+    </div>
+    {/* The borderColor variable can change the color used for the border. */}
+    <div style={{ padding: '1rem', display: 'inline-block' }}>
+      <Icon
+        name="calendar"
+        bordered
+        variables={{ color: 'cornflowerblue', borderColor: 'violet' }}
+      />
+      <Icon
+        name="call"
+        bordered
+        variables={{ color: 'cornflowerblue', borderColor: 'yellowgreen' }}
+      />
+      <Icon
+        name="call-video"
+        bordered
+        variables={{ color: 'cornflowerblue', borderColor: 'orangered' }}
+      />
     </div>
   </div>
 )